io_bench/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==3.7.1
asks==3.0.0
async-generator==1.10
asyncio==4.0.0
attrs==25.4.0
certifi==2025.11.12
contourpy==1.3.3
curio-compat==1.6.7
cycler==0.12.1
fonttools==4.61.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
multidict==6.7.0
numpy==2.4.1
objprint==0.3.0
outcome==1.3.0.post0
packaging==25.0
pillow==12.1.0
propcache==0.4.1
pyinstrument==5.1.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.5.4
trio==0.32.0
typing_extensions==4.15.0
uvloop==0.22.1
winloop==0.4.1
yarl==1.22.0
